move deserialization / conversion code from tests project to business logic class library
