+++ Timetable/Deserialized/Block.cs
@@ -1,0 +1,9 @@
+﻿namespace Timetable.Deserialized
+{
+    public class Block
+    {
+        public string Name { get; set; }
+
+        public string Begin { get; set; }
+    }
+}
