+++ Timetable/Deserialized/Table.cs
@@ -1,0 +1,13 @@
+﻿using Timetable;
+
+namespace Timetable.Deserialized
+{
+    public class Table
+    {
+        public string[] BlockStartTimes { get; set; }
+
+        public BlockStartTime StartTimes => new BlockStartTime(BlockStartTimes);
+
+        public Weekday[] Weekdays { get; set; }
+    }
+}
