+++ Timetable/Deserialized/Weekday.cs
@@ -1,0 +1,9 @@
+﻿namespace Timetable.Deserialized
+{
+    public class Weekday
+    {
+        public string Name { get; set; }
+
+        public Block[] Blocks { get; set; }
+    }
+}
