+++ Timetable/Timetable.csproj
@@ -71,6 +71,9 @@
     <Compile Include="BreakFactory.cs" />
     <Compile Include="BreakStartTime.cs" />
     <Compile Include="DayFactory.cs" />
+    <Compile Include="Deserialized\Block.cs" />
+    <Compile Include="Deserialized\Table.cs" />
+    <Compile Include="Deserialized\Weekday.cs" />
     <Compile Include="DoubleBlock.cs" />
     <Compile Include="Ensure.cs" />
     <Compile Include="Ics.cs" />
@@ -83,6 +86,7 @@
     <Compile Include="Subject.cs" />
     <Compile Include="Table.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TableConvert.cs" />
     <Compile Include="ToniTable.cs" />
     <Compile Include="Week.cs" />
     <Compile Include="Weekday.cs" />
